Refactor error payload logic
